Don't rely on files being .pdi or images when they're copied.
Further work on splitting the graphics/toast part from this library. Previously, we relied on that the images where already loaded anyway. Now we have to copy them without that. An alternative would've been to (still) use the datastore-functionality, since that doesn't pull in graphics either, but in practice this would only save a few lines in the copy function, and make it less usable in the future (since the datastore can _only_ read/write pdi and json and now we already have the facilities to just copy over _anything_).
The only downside to this is that the lua documentation specifies that 'write' saves a string, while we just dump the buffer in, which seems to work for our purposes, but feels a bit like a rug-pull might be possible at some point.