Resolve discrepancy in advance function naming
The `achievements.lua` module uses `advance` and `advanceTo`.
The `toasts.lua` module used `advanceBy` and `advanceTo`. The latter
was setting up wrappers of the former in order to call the appropriate
advance function when configured to show toasts automatically on
advance. Not only was the naming inconsistent, but the wrapper was
attpenting to wrap the `achievements.advanceBy` function, which is nil.

This updates `toasts.lua` to use `advance` and `advanceTo`, bringing
the two into alignment.